+++ aredis/__init__.py
@@ -15,7 +15,7 @@ from aredis.exceptions import (
 )
 
 
-__version__ = '1.0.8'
+__version__ = '1.0.9'
 
 VERSION = tuple(map(int, __version__.split('.')))